*refactoring of ConvertSoftmaxNode.
